refactor(core): consolidate shared read helpers into read_batches (Slice 4, narrow)
Core consolidation Slice 4 (narrow option from docs/core_consolidation_plan.md).
Inline core/read_common.py (shared low-level read retry/error helpers) into
core/read_batches.py with byte-identical function bodies and merged imports.
read_common.py is deleted; core/ file count drops 21 -> 20.

Moved helpers: ILLEGAL_DATA_ADDRESS, is_illegal_data_address_response,
is_transient_error_response, execute_read_call, log_read_retry,
raise_for_error_response.

read_common and read_batches are independent leaves (neither imports the other or
any other read-cluster module), and in production both are consumed only by
core/io_mixin.py (itself imported only by core/client.py). Merging two siblings
that share a single consumer introduces no new import edge and no cycle, so this is
a pure move with zero read/update-cycle behavior change.

Import sites updated directly (no shims):
- core/io_mixin.py (5 `from .read_common import` -> `from .read_batches import`)
- tests/test_read_common.py

Only the narrow plan option was taken. read_bits.py, runtime_io.py, io_mixin.py and
the broader read/ sub-package were NOT touched and remain deferred until real-device
read-path validation reaches PASS. The maintainer explicitly waived the read-path
gate for this zero-behavior-change move.

Docs: docs/core_consolidation_plan.md marks Slice 4 (narrow) DONE with rollback plan
and supersedes the B4 BLOCKED note; docs/coordinator_proxy_remaining_plan.md updated
(read_common -> read_batches); CHANGELOG.md [Unreleased] Internal entry added.

No Modbus register addresses/names, entity IDs, unique IDs, service IDs, or
translation keys changed. Read path behavior unchanged.
